requirements.txt
langchain
google.generativeai
streamlit